refactor: Remove agent.rs; Rename agent_v2/agent.rs -> agent/core.rs
