+++ packages/talos_agent/src/agent/mod.rs
@@ -1,4 +1,4 @@
-pub mod agent_v2;
+pub mod core;
 mod decision_reader;
 pub mod errors;
 pub mod model;
