+++ packages/talos_agent/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod agent;
-pub mod agentv2;
 pub mod api;
 mod messaging;
 pub mod metrics;
